requirements.txt
pycryptodome
pyinstaller